+++ NebulaModel/DataStructures/PlayerData.cs
@@ -8,7 +8,7 @@ namespace NebulaModel.DataStructures
         public string Username { get; set; }
         public ushort PlayerId { get; set; }
         public int LocalPlanetId { get; set; }
-        public Float3 MechaColor { get; set; }
+        public Float4[] MechaColors { get; set; }
         public Float3 LocalPlanetPosition { get; set; }
         public Double3 UPosition { get; set; }
         public Float3 Rotation { get; set; }
@@ -17,13 +17,13 @@ namespace NebulaModel.DataStructures
         public int LocalStarId { get; set; }
 
         public PlayerData() { }
-        public PlayerData(ushort playerId, int localPlanetId, Float3 mechaColor, string username = null, Float3 localPlanetPosition = new Float3(), Double3 position = new Double3(), Float3 rotation = new Float3(), Float3 bodyRotation = new Float3())
+        public PlayerData(ushort playerId, int localPlanetId, Float4[] mechaColors, string username = null, Float3 localPlanetPosition = new Float3(), Double3 position = new Double3(), Float3 rotation = new Float3(), Float3 bodyRotation = new Float3())
         {
             PlayerId = playerId;
             LocalPlanetId = localPlanetId;
+            MechaColors = mechaColors;
             Username = !string.IsNullOrWhiteSpace(username) ? username : $"Player {playerId}";
             LocalPlanetPosition = localPlanetPosition;
-            MechaColor = mechaColor;
             UPosition = position;
             Rotation = rotation;
             BodyRotation = bodyRotation;
@@ -35,7 +35,11 @@ namespace NebulaModel.DataStructures
             writer.Put(Username);
             writer.Put(PlayerId);
             writer.Put(LocalPlanetId);
-            MechaColor.Serialize(writer);
+            writer.Put(MechaColors?.Length ?? 0);
+            for (int i = 0; i < (MechaColors?.Length ?? 0); i++)
+            {
+                MechaColors[i].Serialize(writer);
+            }
             LocalPlanetPosition.Serialize(writer);
             UPosition.Serialize(writer);
             Rotation.Serialize(writer);
@@ -48,7 +52,11 @@ namespace NebulaModel.DataStructures
             Username = reader.GetString();
             PlayerId = reader.GetUShort();
             LocalPlanetId = reader.GetInt();
-            MechaColor = reader.GetFloat3();
+            MechaColors = new Float4[reader.GetInt()];
+            for (int i = 0; i < MechaColors.Length; i++)
+            {
+                MechaColors[i] = reader.GetFloat4();
+            }
             LocalPlanetPosition = reader.GetFloat3();
             UPosition = reader.GetDouble3();
             Rotation = reader.GetFloat3();
@@ -59,7 +67,7 @@ namespace NebulaModel.DataStructures
 
         public IPlayerData CreateCopyWithoutMechaData()
         {
-            return new PlayerData(PlayerId, LocalPlanetId, MechaColor, Username, LocalPlanetPosition, UPosition, Rotation, BodyRotation);
+            return new PlayerData(PlayerId, LocalPlanetId, MechaColors, Username, LocalPlanetPosition, UPosition, Rotation, BodyRotation);
         }
     }
 }