+++ NebulaAPI/CHANGELOG.md
@@ -1,5 +1,10 @@
 ## Changelog
 
+1.1.0:
+
+- Float3 IPlayerData.MechaColor has been changed to Float4[] IPlayerData.MechaColors in line with changes introduced in DSP version 0.8.21.8562.
+- Edited description.
+
 1.0.0:
 
 - initial release on thunderstore